Move NodeCache and CachedNode to the correct package
